alteração double

+++ Calculadora/Program.cs
@@ -6,14 +6,14 @@ namespace Calculadora
 
         static void Main(string[] args)
         {
-            float numero1, numero2;
+            double numero1, numero2;
             
             Console.WriteLine("Digite o primeiro numero: ");
-            numero1 = float.Parse(Console.ReadLine());
+            numero1 = double.Parse(Console.ReadLine());
             Console.WriteLine("Digite o segundo numero: ");
-            numero2 = float.Parse(Console.ReadLine());
-            float subtracao = numero1 - numero2;    
-            float soma = numero1 + numero2;
+            numero2 = double.Parse(Console.ReadLine());
+            double subtracao = numero1 - numero2;
+            double soma = numero1 + numero2;
             Console.WriteLine("O resultado final da soma é: " + soma);
             Console.WriteLine("A subtração dos numero é:" + subtracao);
         }
